requirement.txt
behave==1.2.6
behave-html-formatter==0.9.8
playwright==1.25.2
python-dotenv==1.0.0
